# frontend/lib/api/openapi.json
{
  "components": {
    "schemas": {
      "CreateWebsiteRequest": {
        "description": "Body of `POST /websites`.",
        "properties": {
          "enrich_with_llm": {
            "default": false,
            "description": "Whether this website's runs ask for model-assisted summarization — a written title and description per page, in place of what is extracted from the page's own markup. This records INTENT, not capability: a run actually enriches only when this is true AND model-assisted summarization is enabled for this deployment (an operator-level setting this API never exposes — see ARCHITECTURE.md §11). A run that asked for enrichment and could not get it still completes, using extracted titles and descriptions, and records why in that run's stats. Mechanism and cost: one Claude Haiku 4.5 call per page, on the first 4,000 characters of its extracted text — on the order of cents for a 100-page run.",
            "title": "Enrich With Llm",
            "type": "boolean"
          },
          "url": {
            "description": "An absolute http:// or https:// URL. Stored as submitted; its normalized origin (lowercased scheme and host, default port and path removed) is what deduplicates it against your existing websites.",
            "examples": [
              "https://example.com/docs"
            ],
            "maxLength": 2048,
            "title": "Url",
            "type": "string"
          }
        },
        "required": [
          "url"
        ],
        "title": "CreateWebsiteRequest",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "InstallationResponse": {
        "description": "One GitHub App installation this user has connected.",
        "properties": {
          "account_login": {
            "description": "The GitHub account or organization it was installed on.",
            "title": "Account Login",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "installation_id": {
            "description": "GitHub's own installation id. Not a secret — it grants nothing without this deployment's App private key.",
            "title": "Installation Id",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "installation_id",
          "account_login",
          "created_at"
        ],
        "title": "InstallationResponse",
        "type": "object"
      },
...
      "RepositoryListResponse": {
        "description": "`GET /github/installations/{id}/repositories`.",
        "properties": {
          "repositories": {
            "items": {
              "$ref": "#/components/schemas/RepositoryResponse"
            },
            "title": "Repositories",
            "type": "array"
          },
          "truncated": {
            "description": "Whether the installation has more repositories than this list carries (`MAX_REPOSITORIES`). True means enter the repository by name rather than pick it.",
            "title": "Truncated",
            "type": "boolean"
          }
        },
        "required": [
          "repositories",
          "truncated"
        ],
        "title": "RepositoryListResponse",
        "type": "object"
      },
...
      "WebsiteListItemResponse": {
        "description": "One row of `GET /websites`.\n\n`latest_run` and `schedule` are **always `null` unless `?include=latest_run` was\nrequested**. With it, each is `null` only when the website genuinely has no run / no\nschedule. The two cases are deliberately not distinguished in the payload: a caller that\ndid not ask for the fold has no reason to tell \"not requested\" from \"none exists\", and a\ncaller that did ask knows it asked.",
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "enrich_with_llm": {
            "title": "Enrich With Llm",
            "type": "boolean"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "latest_run": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/LatestRunSummary"
              },
              {
                "type": "null"
              }
            ]
          },
          "origin": {
            "title": "Origin",
            "type": "string"
          },
          "schedule": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ScheduleSummary"
              },
              {
                "type": "null"
              }
            ]
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "url": {
            "title": "Url",
            "type": "string"
          },
          "user_id": {
            "format": "uuid",
            "title": "User Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "user_id",
          "url",
          "origin",
          "title",
          "enrich_with_llm",
          "created_at"
        ],
        "title": "WebsiteListItemResponse",
        "type": "object"
      },
      "WebsiteResponse": {
        "description": "One website, as every endpoint in this feature returns it.\n\nCarries `user_id` because the UI renders the owner and gates the delete control on it —\nand because `require_owner` (`app.core.auth.ownership`) takes the resource, not a bare\nid, so this field is what makes an instance of this model checkable. Exposing it to\nevery signed-in caller is intended: this application is not multi-tenant, and who added\na site is public information within it (ARCHITECTURE.md §4).",
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "enrich_with_llm": {
            "title": "Enrich With Llm",
            "type": "boolean"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "origin": {
            "title": "Origin",
            "type": "string"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "url": {
            "title": "Url",
            "type": "string"
          },
          "user_id": {
            "format": "uuid",
            "title": "User Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "user_id",
          "url",
          "origin",
          "title",
          "enrich_with_llm",
          "created_at"
        ],
        "title": "WebsiteResponse",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "description": "Supabase access token, verified locally against the project's JWKS.",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Backend for llms-text: website registration, crawl runs, llms.txt generation",
    "title": "llms-text API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/github/installations": {
      "get": {
        "description": "Every GitHub App installation the caller has connected.\n\nPer-user, unlike every other read in this API. An installation is a fact about someone's\nGitHub account rather than about a crawled site, which is the distinction\n`internals/publish_reader.py`'s docstring draws against §4.1.",
        "operationId": "list_installations_github_installations_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/InstallationResponse"
                  },
                  "title": "Response List Installations Github Installations Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Installations",
        "tags": [
          "publish"
        ]
      },
      "post": {
        "description": "Record an installation the caller was redirected back from.\n\n`installation_id` is GitHub's own numeric id, arriving from the setup callback's query string.\nThe service verifies it against GitHub before writing anything — see\n`PublishService.connect_installation`, because a number on a redirect is not a fact.\n\nAlways `201`, including when the row already existed and was refreshed. Distinguishing the two\nwould need a branch in this handler, which is what CLAUDE.md #6 forbids; the same reasoning\n`upsert_schedule` gives for always returning `200`.",
        "operationId": "connect_installation_github_installations_post",
        "parameters": [
          {
            "in": "query",
            "name": "installation_id",
            "required": true,
            "schema": {
              "title": "Installation Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InstallationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Connect Installation",
        "tags": [
          "publish"
        ]
      }
    },
    "/github/installations/{id}": {
      "delete": {
        "description": "Forget an installation and every publish target that used it.\n\nDoes not uninstall the App on GitHub — that is the user's own action to take in their account,\nand `PublishService.disconnect_installation` explains why this API does not attempt it.",
        "operationId": "disconnect_installation_github_installations__id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Disconnect Installation",
        "tags": [
          "publish"
        ]
      }
    },
    "/github/installations/{id}/repositories": {
      "get": {
        "description": "The repositories an installation may write to, for the target picker.\n\nProxied live from GitHub rather than stored: the set changes whenever the user edits the\ninstallation's repository access, and a cached copy would offer a repository we can no longer\nwrite to. `502` when GitHub cannot be reached — the request was fine, the upstream was not.",
        "operationId": "list_installation_repositories_github_installations__id__repositories_get",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RepositoryListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Installation Repositories",
        "tags": [
          "publish"
        ]
      }
    },
...
    "/websites": {
      "get": {
        "description": "List **every** website, from every user, most recently active first.\n\nNot filtered by the caller, on purpose: this application is not multi-tenant and\nbrowsing everyone's sites is the product (ARCHITECTURE.md §4.1). `user_id` is present to\nrequire authentication and is intentionally not passed to the service.",
        "operationId": "list_websites_websites_get",
        "parameters": [
          {
            "description": "Set to `latest_run` to fold each website's most recent run and its schedule summary into the response. Omitted, both fields are null.",
            "in": "query",
            "name": "include",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "const": "latest_run",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Set to `latest_run` to fold each website's most recent run and its schedule summary into the response. Omitted, both fields are null.",
              "title": "Include"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/WebsiteListItemResponse"
                  },
                  "title": "Response List Websites Websites Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Websites",
        "tags": [
          "websites"
        ]
      },
      "post": {
        "description": "Register a website, owned by the caller.\n\nThe URL is validated by `CreateWebsiteRequest` before this runs, so a malformed one is a\n`422` that never reaches the service.",
        "operationId": "create_website_websites_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateWebsiteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WebsiteResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WebsiteAlreadyExistsResponse"
                }
              }
            },
            "description": "You have already added a website with this origin. The response carries the existing website's id so the client can navigate to it."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Website",
        "tags": [
          "websites"
        ]
      }
    },
...
    "/websites/{id}/publications": {
      "get": {
        "description": "A website's publication history, newest first, capped at `MAX_PUBLICATIONS_PAGE`.\n\nNot paginated, and not `Page[...]`: this is a panel on a website's page whose list is almost\nalways length zero or one. See `MAX_PUBLICATIONS_PAGE` for the argument.",
        "operationId": "list_publications_websites__id__publications_get",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PublicationResponse"
                  },
                  "title": "Response List Publications Websites  Id  Publications Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Publications",
        "tags": [
          "publish"
        ]
      }
    },
    "/websites/{id}/publish-target": {
      "delete": {
        "description": "Stop publishing a website the caller owns. Idempotent — `204` even if there was none.",
        "operationId": "delete_publish_target_websites__id__publish_target_delete",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Delete Publish Target",
        "tags": [
          "publish"
        ]
      },
      "get": {
        "description": "Return where a website publishes, or `null`.\n\n`200` with a `null` body when the website has no target — a normal state, not an error, exactly\nas `get_schedule` treats a website with no schedule.",
        "operationId": "get_publish_target_websites__id__publish_target_get",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/PublishTargetResponse"
                    },
                    {
                      "type": "null"
                    }
                  ],
                  "title": "Response Get Publish Target Websites  Id  Publish Target Get"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Publish Target",
        "tags": [
          "publish"
        ]
      },
      "put": {
        "description": "Create or replace where a website the caller owns publishes.\n\n`403` when the caller is not the owner, `404` when `installation_id` names no installation of\ntheirs — the second is a `404` rather than a `403` because telling a caller that an\ninstallation exists but belongs to someone else would leak which GitHub accounts other users\nhave connected.",
        "operationId": "upsert_publish_target_websites__id__publish_target_put",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PublishTargetRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublishTargetResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Upsert Publish Target",
        "tags": [
          "publish"
        ]
      }
    },
...
    "/websites/{id}/schedule": {
      "get": {
        "description": "Return one website's schedule. Any signed-in user may read any website's schedule.\n\n`200` with a `null` body when the website exists but has no schedule yet — that is a\nnormal state, not an error. `404` is reserved for an unknown *website*; a missing\nschedule never produces one.\n\nThe path parameter is `id`, not `website_id`, because the noun is already in the path\n(ARCHITECTURE.md §10.3), matching `get_website` in `api/routers/websites.py`.",
        "operationId": "get_schedule_websites__id__schedule_get",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/ScheduleResponse"
                    },
                    {
                      "type": "null"
                    }
                  ],
                  "title": "Response Get Schedule Websites  Id  Schedule Get"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Schedule",
        "tags": [
          "schedules"
        ]
      },
      "put": {
        "description": "Create or replace the schedule for a website the caller owns.\n\nAlways `200`, never a conditional `201` for the create case. Distinguishing \"this was a\ncreate\" from \"this was an update\" would require a branch in this handler — checking\nwhether a schedule already existed — and CLAUDE.md #6 forbids exactly that: a router\nparses input, calls one service method, and returns. The ticket's own wording (\"200/201\")\npermits either; `200` is the one that keeps this file free of logic.\n\n`403`, not `404`, when the caller is not the website's owner — `GET /websites/{id}`\nalready returns that website to every signed-in user, so hiding its existence here would\nbe inconsistent with the endpoint next to it (ARCHITECTURE.md §4).",
        "operationId": "upsert_schedule_websites__id__schedule_put",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpsertScheduleRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScheduleResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Upsert Schedule",
        "tags": [
          "schedules"
        ]
      }
    },
...
  }
}